+++ pkg/webscan/InputType.go
@@ -1,0 +1,9 @@
+package webscan
+
+type InputType uint8
+
+const (
+	Domain InputType = iota
+	IPv4
+	IPv6
+)