+++ pkg/webscan/PrintDnsScanResults.go
@@ -3,34 +3,26 @@ package webscan
 import (
 	"fmt"
 	"strings"
-
-	"github.com/thetillhoff/webscan/pkg/dnsScan"
 )
 
-func (engine Engine) PrintDnsScanResults() {
+func (engine Engine) PrintDnsScanEngines() {
 	if engine.DetailedDnsScan {
-		dnsServer := engine.DnsScanEngine.GetCustomDns()
-		if dnsServer != "" {
-			fmt.Println("Using custom dns server:", dnsServer)
-		} else {
-			fmt.Println("Using system dns server")
-		}
 
 		fmt.Println()
-		if len(engine.DnsScanEngine.DomainOwners) == 0 {
+		if len(engine.dnsScanEngine.DomainOwners) == 0 {
 			fmt.Println("Could not retrieve Domain Owner (country TLDs are not supported yet by RDAP)")
 		} else {
-			fmt.Println("Domain Registrar: ", strings.Join(engine.DnsScanEngine.DomainOwners, ", "))
+			fmt.Println("Domain Registrar: ", strings.Join(engine.dnsScanEngine.DomainOwners, ", "))
 		}
 
 		fmt.Println()
 		fmt.Println("DNS records:")
-		engine.DnsScanEngine.PrintAllDnsRecords()
+		engine.dnsScanEngine.PrintAllDnsRecords()
 
-		if engine.Opinionated {
-			// Domain Accessibility
+		// Domain Accessibility
+		if len(engine.dnsScanEngine.OpinionatedHints) > 0 {
 			fmt.Println()
-			for _, hint := range dnsScan.GetDomainAccessibilityHints(engine.url) {
+			for _, hint := range engine.dnsScanEngine.OpinionatedHints {
 				fmt.Println(hint)
 			}
 		}