+++ pkg/webscan/ScanMailConfig.go
@@ -1,12 +1,17 @@
 package webscan
 
-import "errors"
+import (
+	"errors"
+	"fmt"
+)
 
-func (engine Engine) ScanMailConfig() (Engine, error) {
+func (engine Engine) ScanMailConfig(inputUrl string) (Engine, error) {
+
+	fmt.Println("Scanning mail config...")
 
 	if engine.SubdomainScan {
 		if engine.DkimSelector != "" {
-			engine.mailConfigRecommendations = engine.DnsScanEngine.CheckMailSecurity(engine.url, engine.DkimSelector)
+			engine.mailConfigRecommendations = engine.dnsScanEngine.CheckMailSecurity(inputUrl, engine.DkimSelector)
 		} else {
 			return engine, errors.New("DKIM selector required")
 		}