+++ pkg/webscan/ScanPortSimple.go
@@ -1,6 +1,10 @@
 package webscan
 
-import "github.com/thetillhoff/webscan/pkg/portScan"
+import (
+	"fmt"
+
+	"github.com/thetillhoff/webscan/pkg/portScan"
+)
 
 func (engine Engine) ScanPortSimple() (Engine, error) {
 	var (
@@ -12,7 +16,9 @@ func (engine Engine) ScanPortSimple() (Engine, error) {
 		openPortsPerIp map[string][]uint16
 	)
 
-	openPortsPerIp = portScan.ScanPortRangeOfIps(append(engine.DnsScanEngine.ARecords, engine.dnsScanResult.AAAARecords...), scanPorts)
+	fmt.Println("Scanning ports (simple)...")
+
+	openPortsPerIp = portScan.ScanPortRangeOfIps(append(engine.dnsScanEngine.ARecords, engine.dnsScanEngine.AAAARecords...), scanPorts, engine.Verbose)
 
 	engine.portScanOpenPorts, engine.portScanInconsistencies = portScan.CompareOpenPortsOfIps(openPortsPerIp)